split up index.sx
