+++ package.json
@@ -38,7 +38,6 @@
     "is-hotkey": "^0.1.6",
     "nanoid": "^3.1.12",
     "react": "^16.13.1",
-    "react-contenteditable": "^3.3.5",
     "react-dom": "^16.13.1",
     "slate": "^0.59.0",
     "slate-react": "^0.59.0",
