+++ src/components/Button.tsx
@@ -1,10 +1,8 @@
-import React from 'react';
-
 type Props = {
   onClick: () => void;
 };
 
-const Button: React.FunctionComponent<Props> = (props) => {
+const Button = (props: Props) => {
   const { children, onClick } = props;
   return (
     <button
